fix use of Future and changed default keep alive value to a positive value as negative values are not allowed anymore.

+++ src/org/jgroups/jmx/protocols/TCP_NIOMBean.java
@@ -2,7 +2,7 @@ package org.jgroups.jmx.protocols;
 
 /**
  * @author Scott Marlow
- * @version $Id: TCP_NIOMBean.java,v 1.1 2005/11/22 13:58:18 smarlownovell Exp $
+ * @version $Id: TCP_NIOMBean.java,v 1.2 2007/01/10 19:38:18 smarlownovell Exp $
  */
 public interface TCP_NIOMBean extends TCPMBean {
 
@@ -12,5 +12,5 @@ public interface TCP_NIOMBean extends TCPMBean {
    int getProcessorMinThreads();
    int getProcessorMaxThreads();
    int getProcessorQueueSize();
-   int getProcessorKeepAliveTime();
+   long getProcessorKeepAliveTime();
 }
